Add no preload error to versioned library.
Previously, only the loader would complain if it was dynamically
loaded after preload. This PR adds a FATAL error message if the
versioned .so is not preloaded AND the loader has not been previously
loaded. This is a common case if the user forgets to edit
shared_preload_libraries to add the loader.

This PR also fixes the error in cases when it's encountered by a
non-superuser.